fix: core — 20 compilation errors across 5 root causes

P0-1: OvershootHelper.resetOvershoot() — moved from inside object to
  top-level extension (was inaccessible to LegitAimStrategy/SelfAdaptive).

P0-2: AimStrategy.State — made 'open class' so AdaptiveState can extend.
  AdaptiveState now overrides reset() to clear alignmentEma etc.

P1-1: ProbabilisticClickStrategy — added missing import ClickMode.

P1-2: LegitVelocityStrategy.processLegit — added state parameter
  (used state.delayQueue/state.tickCounter but never received state).

P2: GaussianNoise.ln(s) → ln(s.toDouble()), RotationCalculator
  bestAngularDist = angularDist.toDouble() (Float→Double mismatch).

+++ SwitchLite/core/src/main/kotlin/io/switchlite/core/algorithm/GaussianNoise.kt
@@ -37,7 +37,7 @@ class GaussianNoise(
             s = u * u + v * v
         } while (s >= 1 || s == 0f)
         
-        val mul = sqrt(-2.0 * ln(s) / s)
+        val mul = sqrt(-2.0 * ln(s.toDouble()) / s)
         val z0 = u * mul
         val z1 = v * mul
